Backend/requirements.txt
aniso8601==10.0.0
annotated-types==0.7.0
asn1crypto==1.5.1
attrs==25.3.0
bcrypt==4.3.0
bidict==0.23.1
blinker==1.9.0
boto3==1.38.13
botocore==1.38.13
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
dnspython==2.7.0
eventlet==0.33.3
Flask==3.1.0
Flask-JWT-Extended==4.7.1
Flask-Login==0.6.3
flask-restx==1.3.0
Flask-SocketIO==5.5.1
greenlet==3.2.2
h11==0.14.0
idna==3.10
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
mypy==1.15.0
mypy_extensions==1.1.0
oauthlib==3.2.2
peewee==3.17.9
pg8000==1.31.2
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-engineio==4.12.0
python-socketio==5.13.0
pytz==2025.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
s3transfer==0.12.0
scramp==1.4.5
simple-websocket==1.1.0
six==1.17.0
tomli==2.2.1
typing-inspection==0.4.0
typing_extensions==4.13.1
urllib3==1.26.20
Werkzeug==3.1.3
wsproto==1.2.0
